refs #2

+++ .travis/build.sh
@@ -28,18 +28,22 @@ EOF
     docker cp         "tini:/usr/src/tini/tini-static" "$PROJECT/tini"
 }
 
-docker pull   "$REPOSITORY/openjdk-$ARCH:8-jdk"
-docker tag -f "$REPOSITORY/openjdk-$ARCH:8-jdk" "java:8-jdk"
+[[ "$JAVA" =~ ^8-jdk.* ]];
+
+docker pull   "$REPOSITORY/java-$ARCH:$JAVA"
+docker tag -f "$REPOSITORY/java-$ARCH:$JAVA" "java:8-jdk"
 
 case "$CUSTOM" in
-    walle )
+    walle* )
         docker build -t "$TAG:$TAGSPECIFIER"      \
                      --build-arg SCRIPT="$SCRIPT" \
-                     "contrib/$CUSTOM"
+                     "contrib/walle"
 
         docker run --rm "$TAG:$TAGSPECIFIER" -v
         ;;
     * )
+        if [ "$JAVA" = "8-jdk" ]; then sed -i -r '/ENTRYPOINT/ s!/bin/tini!docker-eula-java", "/bin/tini!g' "$PROJECT/Dockerfile"; fi
+
         patch -p1 --no-backup-if-mismatch --directory=$PROJECT < .patch/Dockerfile.patch
 
         mktini
@@ -49,6 +53,12 @@ case "$CUSTOM" in
                      --build-arg JENKINS_SHA="$SHA"         \
                      "$PROJECT"
 
-        docker run --rm "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version
+        if [ "$JAVA" = "8-jdk" ]; then
+             if docker run --rm -e eula-java=accept "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version; then             true; fi
+             if docker run --rm -e eula-java=       "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version; then false; else true; fi
+        else
+                docker run --rm                     "$TAG:$TAGSPECIFIER" java -jar /usr/share/jenkins/jenkins.war --version
+        fi
+
         ;;
 esac
